build(scripts): make SSR optional
separate 'start' and 'build' with SSR

improves #30
