+++ config/webpack-config/client/client.dev.js
@@ -2,6 +2,7 @@ const webpack = require('webpack');
 const WriteFileWebpackPlugin = require('write-file-webpack-plugin');
 
 const baseConfig = require('./client.base');
+
 const generateSourceMap = process.env.OMIT_SOURCEMAP === 'true' ? false : true;
 
 const config = {
